+++ src/components/Forms/Common/overview.common.tsx
@@ -22,7 +22,7 @@ import {
   FORM_COMMON_LAYOUT_EXPAND_OFFSET,
   FORM_COMMON_LAYOUT_MARGIN_TOP,
 } from './layout.constant';
-import OverviewShareComponent from './overview.share.component';
+import OverviewShareComponent from './OverviewShareComponent';
 
 interface Props {
   ind?: number;
